Safari-609.1.11.1/LayoutTests/platform/mac-highsierra-wk1/imported/w3c/web-platform-tests/mathml/relations/html5-tree/tabindex-002-expected.txt


FAIL Elements with different tabindex must be focused sequentially when pressing 'Tab' keys assert_array_equals: property 6, expected "text6" but got "text11"
tabindex(omitted)
tabindex(empty)
tabindex(a)
tabindex(-1)
tabindex(0)
tabindex(href)
tabindex(3)
tabindex(2)
tabindex(2)
tabindex(2)
tabindex(1)
